requirements.txt
streamlit
svgwrite
requests
fastapi
uvicorn
python-dotenv
google-generativeai
openai
cairosvg==2.7.1
cachetools
pydantic>=2.0.0
